requirements.txt
python-binance
pandas
numpy
tensorflow
scikit-learn
pyyaml
python-dotenv
SQLAlchemy
ta
matplotlib
